Add files via upload
Download the zip and then unzip it, or make a new folder and save the three .m files in that folder. After that just open the ConvolutionModeSelectorApp.m file in matlab and run the file. It will enable the user to convolute any signals, whether it's DT or CT. Even this app will give the user an opportunity to animate the whole convolution visually to make understand convolution properly. Also the correlation between two signals can be shown.

MATLAB R2023b or newer versions can be used to run the files.